requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
jinja2>=3.1
httpx>=0.28
python-dotenv>=1.0
